Import internal commit 136c4b5ac15a30254a505e6e2f32ec52cd105e20 : [APPROVAL] pending/master/fix-warnings-issues-41_42_46_47_56-drqs-37220936 to master

+++ groups/bsl/bslstl/bslstl_list.t.cpp
@@ -4439,7 +4439,7 @@ void TestDriver<TYPE,ALLOC>::testInsert()
             TestEnum n = TWO, v = NINETYNINE;
 
             x.insert(X.begin(), n, v);
-            ASSERT(X.size()  == n);
+            ASSERT(X.size()  == (size_t)n);
             ASSERT(X.front() == v);
             ASSERT(X.back()  == v);
         }
@@ -6223,7 +6223,7 @@ void TestDriver<TYPE,ALLOC>::testConstructor()
             list<IntWrapper, ALLOC> x(n, v);
             list<IntWrapper, ALLOC>& X = x;
 
-            ASSERT(X.size()  == n);
+            ASSERT(X.size()  == (size_t)n);
             ASSERT(X.front() == v);
             ASSERT(X.back()  == v);
         }
